Add support for shuffle (#3)
This commit adds support for theme shuffles.

To use, pass the first node for the theme into the install_shuffle method and it'll automatically install based on what's selected. 

Note that right now, *all themes in /Themes/ are selected for debug purposes.* This means if you have more than 10 themes in /Themes/ it will not work right now. This will be changed once an interface for selecting themes is done.